Addresses issue-7554 w/ refactor of netty utils to cookie encoding
