+++ framework/src/play-cookie-encoding/src/main/java/play/core/cookie/encoding/ClientCookieEncoder.java
@@ -13,11 +13,12 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package play.core.netty.utils;
+package play.core.cookie.encoding;
 
-import static play.core.netty.utils.CookieUtil.*;
+import static play.core.cookie.encoding.CookieUtil.*;
 
 import java.util.Iterator;
+
 
 /**
  * A <a href="https://tools.ietf.org/html/rfc6265">RFC6265</a> compliant cookie encoder to be used client side,
