+++ framework/src/play-cookie-encoding/src/main/java/play/core/cookie/encoding/ServerCookieEncoder.java
@@ -13,9 +13,9 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package play.core.netty.utils;
+package play.core.cookie.encoding;
 
-import static play.core.netty.utils.CookieUtil.*;
+import static play.core.cookie.encoding.CookieUtil.*;
 
 import java.util.ArrayList;
 import java.util.Collection;
