+++ framework/src/play/src/test/scala/play/api/http/HttpConfigurationSpec.scala
@@ -7,9 +7,10 @@ import java.io.File
 
 import com.typesafe.config.ConfigFactory
 import org.specs2.mutable.Specification
-import play.api.{ Configuration, Environment, Mode, PlayException }
+import play.api.{Configuration, Environment, Mode, PlayException}
 import play.api.mvc.Cookie.SameSite
-import play.core.netty.utils.{ ClientCookieDecoder, ClientCookieEncoder, ServerCookieDecoder, ServerCookieEncoder }
+import play.core.cookie.encoding.{ClientCookieDecoder, ClientCookieEncoder, ServerCookieDecoder, ServerCookieEncoder}
+import play.core.cookie.encoding.{ClientCookieDecoder, ServerCookieDecoder, ServerCookieEncoder}
 
 class HttpConfigurationSpec extends Specification {
