+++ framework/src/play/src/test/scala/play/api/mvc/CookiesSpec.scala
@@ -3,11 +3,12 @@
  */
 package play.api.mvc
 
-import java.time.{ Instant, ZoneId }
+import java.time.{Instant, ZoneId}
 
 import org.specs2.mutable._
-import play.api.http.{ JWTConfiguration, SecretConfiguration, SessionConfiguration }
+import play.api.http.{JWTConfiguration, SecretConfiguration, SessionConfiguration}
 import play.api.mvc.Cookie.SameSite
+import play.core.cookie.encoding.ServerCookieEncoder
 import play.core.test._
 
 import scala.concurrent.duration._
@@ -44,7 +45,7 @@ class CookiesSpec extends Specification {
 
   "ServerCookieEncoder" should {
 
-    val encoder = play.core.netty.utils.ServerCookieEncoder.STRICT
+    val encoder = ServerCookieEncoder.STRICT
 
     "properly encode ! character" in {
       val output = encoder.encode("TestCookie", "!")
